Changed how mouse selector works
Changed it so that instead of clicking on the mouse to change player to the mouse, the player instead clicks on the mouse's sprite. This was done to prevent a bug where the player could select crates while the mouse was standing on one, and also to increase usability as the box for selecting the mouse is now larger.